backend/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
Pillow>=10.0.0
numpy>=1.26.0
